Fix - Move movi-card to shared components folder
