+++ src/app/app.module.ts
@@ -1,12 +1,6 @@
 import { ListComponent } from './shared/components/list/list.component';
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
-import { SearchBarComponent } from './search-bar/search-bar.component';
-import { NominationsContainerComponent } from './nominations-container/nominations-container.component';
-import { ResultsComponent } from './results/results.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { FormsModule } from '@angular/forms';
@@ -14,7 +8,13 @@ import { MatInputModule } from '@angular/material/input';
 import { MatIconModule } from '@angular/material/icon';
 import { HttpClientModule } from '@angular/common/http';
 import { MatButtonModule } from '@angular/material/button';
-import { MovieCardComponent } from './results/movie-card/movie-card.component';
+
+import { AppRoutingModule } from './app-routing.module';
+import { AppComponent } from './app.component';
+import { SearchBarComponent } from './search-bar/search-bar.component';
+import { NominationsContainerComponent } from './nominations-container/nominations-container.component';
+import { ResultsComponent } from './results/results.component';
+import { MovieCardComponent } from './shared/components/movie-card/movie-card.component';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { LoaderComponent } from './shared/components/loader/loader.component';
 import { BannerComponent } from './shared/components/banner/banner.component';
