--- src/app/results/movie-card/movie-card.component.html
@@ -1,17 +1,0 @@
-<div @items class="card">
-    <ng-container
-        *ngIf="movieInformation.Poster && movieInformation.Poster.toLocaleUpperCase() !== 'N/A'; else brokenImg">
-        <img [src]="movieInformation.Poster" alt="No poster found">
-    </ng-container>
-    <ng-template #brokenImg>
-        <mat-icon class="card__broken-img">broken_image</mat-icon>
-    </ng-template>
-    <div class="card__desc">
-        <span class="card__title">{{movieInformation.Title}}</span>
-        <span><i>{{movieInformation.Type}}</i></span>
-        <span>({{movieInformation.Year}})</span>
-    </div>
-    <button mat-raised-button color="primary" [disabled]="isDisabled" (click)="onButtonClick($event)">
-        {{ isDisabled ? getDisabledLabel() : primaryBtnActiveLabel }}
-    </button>
-</div>
